Split runtime/runner.py (5,245 lines) into focused modules
The monolithic runner.py has been refactored into 8 focused submodules
for better maintainability:

- _base.py (~286 lines): Base runner class and progress callback utilities
- _determinism.py (~537 lines): Deterministic run ID and timestamp generation
- _result_utils.py (~879 lines): Result conversion and spec building utilities
- _artifact_utils.py (~581 lines): Run directory and JSONL handling utilities
- _config_loader.py (~600 lines): Configuration loading and object creation
- _sync_runner.py (~731 lines): Synchronous ProbeRunner implementation
- _async_runner.py (~695 lines): Asynchronous AsyncProbeRunner implementation
- _high_level.py (~896 lines): High-level config-driven experiment functions

The runner.py module is now a thin re-export wrapper (~180 lines) that
maintains full backward compatibility by re-exporting all public APIs.

TASK 2 of 4 in the enterprise-grade refactoring plan.
